FEM: move selection widget into PyGui
